ブックマーク機能完成

+++ app/controllers/bookmarks_controller.rb
@@ -7,14 +7,19 @@ class BookmarksController < ApplicationController
 
   def create
     post = Post.find(params[:post_id])
-    current_user.bookmarks.create(post: post)
-    redirect_to posts_path, notice: 'ブックマークしました'
+    current_user.bookmarks.find_or_create_by(post: post)
+    redirect_to request.referer || posts_path, notice: 'ブックマークしました'
   end
 
   def destroy
     post = Post.find(params[:post_id])
     bookmark = current_user.bookmarks.find_by(post: post)
-    bookmark.destroy
-    redirect_to posts_path, notice: 'ブックマークを取り消しました'
+    
+    if bookmark
+      bookmark.destroy
+      redirect_to request.referer || posts_path, notice: 'ブックマークを取り消しました'
+    else
+      redirect_to request.referer || posts_path, alert: 'ブックマークが見つかりませんでした'
+    end
   end
 end
